[dev-overlay] Discontinue using `rem` for the UI (#76969)
This PR reverts our decision to use `rem` for the Dev Overlay surfaces. 

We have realised that using `rem` to style the Dev Overlays is not a
good idea because if an app sets their root font size to something tiny,
they will not have a great time using the Overlays, and it feels
unexpected to have the app root size leak into an external package.

I did not remove the references to `--size-*` CSS vars because we still
want to offer a way to scale the size of the UI, just through a custom
interface in the dropdown menu

---
- https://github.com/vercel/next.js/discussions/76812
- Closes NDX-969