[dev-overlay] Customize `<select>` styling for consistency (#76973)
This PR applies `all: unset` to the `<select>` we use in Preferences for
the Dev Overlay and adds a custom icon because we lose the chevron as
soon as we apply this.

Why?

So the styling for the control would look more consistent, for example
this is what it looks like in Safari now:

![CleanShot 2025-03-11 at 12 36
12@2x](https://github.com/user-attachments/assets/a6da93ce-2b89-4420-8165-c58d4ec8ba29)

And after the changes:

![CleanShot 2025-03-11 at 12 36
24@2x](https://github.com/user-attachments/assets/4f0f4c31-cc0f-436b-9575-e4434c783b4e)

---
Closes NDX-928