[dev-overlay] Add `size` setting to preferences (#77027)
This PR is a follow up to https://github.com/vercel/next.js/pull/76969
and adds a size setting to the Dev Tools preferences.

Right now the **Small**, **Medium** (default), and **Large** variants
scale the UI as if the root font size were **14px**, **16px**, **18px**.
We can tweak these values as a follow-up.


https://github.com/user-attachments/assets/43055cae-1e14-4e52-94cc-3ad1b8cae424

---
- Closes NDX-971