Make `revalidate*` work when followed by a redirect in a route handler
When a route handler calls `revalidateTag` or `revalidatePath`, we need
to make sure that these revalidations are executed before returning
with the redirect response.

Note: When deployed, we currently don't have a strong guarantee that the
revalidations are fully propagated before executing the redirect target
route – as we do for redirecting server actions. This is because, for
route handlers, the redirect occurs client-side, which prevents us from
using the same technique as for server actions, which involves sending a
revalidate token as a request header. This token must not leak to the
client. However, eventually the revalidation will be propagated, and a
refresh should show fresh data.

fixes #74757