Slightly improve error handling for unknown server actions (#77135)
This fixes the following error when retrieving a non-existent server
action from the server reference manifest:
`Cannot read properties of undefined (reading 'workers')`.

We're also removing the original error from the error message with this
PR, as there shouldn't be an original error anymore. The result will
just be `undefined`.