[metadata] remove the default segement check for metadata rendering (#77119)
### What

Remove the default segment checks for metadata rendering. When the
`default.js` is rendered, it should still have the metadata. It was a
condition added due to possible of duplicated metadata, but it's
resolved by other conditions such as only rendering in child slot, and
hoisting metadata into rsc head for navigation.