devlow-bench: wait for complete ready for server startup event
Previously, we marked the server startup whenever a local url was shown in stdio. Next.js shows this url as soon as it can to make it actionable sooner, before the server is truly ready.

This changes makes the event record the time until the “Ready in Nms” message is shown instead.

Test Plan: `./node_modules/.bin/devlow-bench ./scripts/devlow-bench.mjs` and validate server startup is recorded.