[Segment Cache] Implement staleTimes.dynamic
When navigating to the same page multiple times in a single session,
we should re-use data from the previous navigation up to the threshold
given by `staleTimes.dynamic`.

I used the `navigatedAt` timestamp added in the previous PR to track
whether a segment is stale.

The default threshold is 0 seconds, but it used to be 30 seconds;
the config option mostly exists so developers can restore the previous
behavior. However, in the future we will likely re-introduce a more
composable version of the same concept, akin to a `cacheLife` that works
for dynamic data.

We will likely also unify this with how back/forward cache works, since
this is another case where dynamic data is reused.

I confirmed that this fixes the tests in client-cache.original.test.ts.