[perf] skip loading client manifest for static metadata routes (#77260)
### What

During development we had a retry logic to load the client manifest for
App Router routes, which is the requirement since #74835. But for static
metadata routes like `favicon.ico` it doesn't generate client manifest,
loading it will cause extra loading time.

Refactoring the metadata helpers to determine the static metadata routes
more strictly, mostly only cover the image one as their pathname and
entry filename is pretty consistent. For sitemap or robots.txt it's
still fine to leave as it in dev for now. The delay is mostly waiting
for manifest.

Long term goal is to build the static metadata routes as static output
rather than compile each time.