Fix type definitions for MDX plugins
`@next/mdx` allows users to pass an array as a plugin where the first
member is a string referencing the plugin. This is not allowed by the
MDX options.

This changes that by allowing a string whose first member is a string.
It’s also more strict about the plugin types that may be passed in. This
prevents users from accidentally passing a remark plugin into
`rehypePlugins`, etc.

Ideally `@next/mdx` should define types dependencies in `package.json`,
but for now I refrained from doing so.