fix: absolute assetPrefix url with path (#77256)
### What?

When an absolute url with path has been set as the assetPrefix, the path
is not used when serving the bundles. Instead the bundles are served
directly on the root. The rewrite did not seem to take a path into
account, only a host.

e.g. when the assetPrefix is set to
`https://example.vercel.sh/custom-asset-prefix`, it does not serve the
bundles from `<host>/custom-asset-prefix/_next/*` instead only from
`<host>/_next/*`

Related; https://github.com/vercel/next.js/discussions/77118

---------

Co-authored-by: JJ Kasper <jj@jjsweb.site>