Move action-queue to client-only directory
This was in a shared directory, but it's only used on the client. I've
moved it to a client-only directory to make it less likely that a
refactor causes server code to leak into the client bundle.

I also renamed it to app-router-instance because I'm about to move
move more router-related code into here.