[metadata] always serve streaming metadata in build (#77437)
### What

During the export phase in next build, we always enable the streaming metadata since if there's any dynamic access in metadata we can determine it in the build phase. If it's static, then it won't affect anything. If it's dynamic, then it can be handled when request hits the route.

### Why

We want it enabled even during build b/c this is how we will enforce DIO rules without making any dynamic metadata automatically a build failure