Remove obsolete update of implicit tags expiration after server action
With the change in #76885, we don't need to update the expiration date
of implicit tags anymore at the end of a request to ensure no stale
data is used after `revalidatePath` was called in a server action.

This PR is covered by the existing tests, e.g. `use-cache should
revalidate caches after redirect`.