Remove useSyncExternalStore from useIsDevRendering
useIsDevRendering is a global pending state for router navigations/
actions. The idiomatic API for this pattern is useOptimistic.

Previously, this was implemented using useSyncExternalStore, which led
to an additional sync render every time a navigation completed. The
advantage of useOptimistic is that it gets automatically reverted in the
same commit as the transition it's associated with.