Turbopack Build: CSR bailout test skip check for file path (#77639)
### What?

Fix Turbopack test for missing suspense with CSR bailout by
conditionally checking for source file path in error output only when
not using Turbopack.

### Why?
The path in the error output is different when using Turbopack, causing
the test to fail. This change makes the test pass in both environments
by skipping the path check when Turbopack is enabled.

Applying sourcemaps during prerendering is being worked on. We'll be
able to un-skip the particular check when it's ready.