legacyBehavior deprecation error should only trigger once (#77687)
We added a deprecation warning for `legacyBehavior` on the `Link`
component but since we show errors in the error overlay, this could lead
to a lot of noise if many links are using the `legacyBehavior` prop.

This copies over the `warn-once` utility to `error-once` and uses that
utility to dedupe.