Ignore an existing HMR refresh hash cookie with `next start`
This is a follow-up fix for #75474.

When switching from `next dev` to `next start` locally, a user might
have an existing HMR refresh hash session cookie.

We need to make sure that the hash value is excluded from any `"use
cache"` cache keys, when running with `next start`. Otherwise, the cache
keys will be different when prerendering vs. handling a dynamic request,
e.g. when executing a server action.

In production, we now also omit the serialized `"$undefined"` value for
the HMR refresh hash from the cache keys to keep them shorter.