Optimize server runtime bundles
Previously, only the prod runtime bundles were split by bundler. In dev
mode, a combined runtime bundle was used. As a consequence, both
`react-server-dom-webpack` and `react-server-dom-turbopack` were
included in this bundle.

This not only impacts the startup time for the dev server, but also
impacts HMR times for Turbopack, because the runtime bundle is deleted
from the require cache for every server component edit.

In addition, even the prod runtime for Turbopack unnecessarily included
`react-server-dom-webpack`, which was caused by incorrect import
aliases.

This PR decreases the dev runtime bundle size for both Webpack and
Turbopack by 16%, and it decreases the prod runtime bundle size for
Turbopack by 12%.