[dynamicIO] only abort once per prerender (#77747)
Prior to this change multiple sync dynamic APIs could trigger multiple
aborts. Generally this isn't a problem however it is wasteful compute
and blocks a later refactor where we stop explicitly tracking whether
the sync dynamic access is what caused the prerender to abort or not. To
achieve better perf and future semantics we now only abort if the
prerender is not yet aborted. In effect this means that there can only
be one sync abort error per prerender in RSC and SSR respectively.

---------

Co-authored-by: Hendrik Liebau <mail@hendrik-liebau.de>