Turbopack build: Move Turbopack marker to SERVER_FILES_MANIFEST (#77711)
### What?

As recommended by @ijjk.

Replaces the `IS_TURBOPACK_BUILD_FILE` marker with a `turbopack` boolean
property in the required server files manifest.

Would like a review of this because it does mean we have to ship the
require server files manifest into the standalone output which seemingly
did not happen before. If we'd want to avoid that the current approach
of writing a small file is better.


### How?

- Added a `turbopack` boolean property to the
`RequiredServerFilesManifest` interface
- Removed the `IS_TURBOPACK_BUILD_FILE` constant and related file
operations
- Updated the server to check the manifest property instead of looking
for the marker file
- Updated tests to verify the manifest property instead of the marker
file