chore(turbo-tasks): Audit all remaining uses of `Vc` in a struct (#77756)
I suspect the test failures in https://github.com/vercel/next.js/pull/77661 are related to a `Vc` that should be an `OperationVc`.

I realized that we have so few uses of `Vc` inside structs and enums that it might be easier to work off of that list than to carefully audit my changes in my other PR, though it doesn't look like anything's a smoking gun here.

Either way, while it's totally valid for non-`VcValueType` structs and enums to contain unresolved `Vc`s, it's usually suboptimal, and there should probably at least be a comment explaining the choice to use an unresolved `Vc`.

I still have a long-term goal of renaming `ResolvedVc` to `Vc` and the current version of `Vc` to `UnresolvedVc`, so for that reason alone, I want everything to prefer using `ResolvedVc` "by default".