Client instrumentation: onRouterTransitionStart
Adds a new API for observing the start of an App Router navigation:

```ts
// <PROJECT_ROOT>/instrumentation-client.ts

export function onRouterTransitionStart(
  url: string,
  navigationType: 'push' | 'replace' | 'traverse'
) {
  // ...
}
```

`navigationType` is one of "push", "replace", or "traverse". This is
inspired by the Navigation API:
https://developer.mozilla.org/en-US/docs/Web/API/NavigateEvent/navigationType

`onRouterTransitionStart` is intended for sending logs to a performance
monitoring tool. It is _not_ intended as a general purpose event for
implementing application behavior.

It fires at the start of every client-side navigation, including those
initiated by a popstate (back/forward) event.

There is no corresponding API for observing the end of a navigation. We
intend to build support for this in the future, but it's a non-trivial
problem space due of the streaming nature of React transitions. Refer
to the React Transition Tracing proposal for more details:
https://github.com/reactjs/rfcs/blob/transition-tracing/text/0235-transition-tracing.md

`onRouterTransitionStart` is only called during App Router navigations.
To instrument Pages Router navigations, use `router.events`.