Client instrumentation: onRouterTransitionStart (#77791)
Adds a new API for observing the start of an App Router navigation:

```ts
// <PROJECT_ROOT>/instrumentation-client.ts

export function onRouterTransitionStart(
  url: string,
  navigationType: 'push' | 'replace' | 'traverse'
) {
  // ...
}
```

`navigationType` is one of "push", "replace", or "traverse". This is
inspired by the Navigation API:

https://developer.mozilla.org/en-US/docs/Web/API/NavigateEvent/navigationType

`onRouterTransitionStart` is intended for sending logs to a performance
monitoring tool. It is _not_ intended as a general purpose event for
implementing application behavior.

It fires at the start of every client-side navigation, including those
initiated by a popstate (back/forward) event.

There is no corresponding API for observing the end of a navigation. We
intend to build support for this in the future, but it's a non-trivial
problem space due of the streaming nature of React transitions. Refer to
the React Transition Tracing proposal for more details:
https://github.com/reactjs/rfcs/blob/transition-tracing/text/0235-transition-tracing.md

`onRouterTransitionStart` is only called during App Router navigations.
To instrument Pages Router navigations, use `router.events`.

<!-- Thanks for opening a PR! Your contribution is much appreciated.
To make sure your PR is handled as smoothly as possible we request that
you follow the checklist sections below.
Choose the right checklist for the change(s) that you're making:

## For Contributors

### Improving Documentation

- Run `pnpm prettier-fix` to fix formatting issues before opening the
PR.
- Read the Docs Contribution Guide to ensure your contribution follows
the docs guidelines:
https://nextjs.org/docs/community/contribution-guide

### Adding or Updating Examples

- The "examples guidelines" are followed from our contributing doc
https://github.com/vercel/next.js/blob/canary/contributing/examples/adding-examples.md
- Make sure the linting passes by running `pnpm build && pnpm lint`. See
https://github.com/vercel/next.js/blob/canary/contributing/repository/linting.md

### Fixing a bug

- Related issues linked using `fixes #number`
- Tests added. See:
https://github.com/vercel/next.js/blob/canary/contributing/core/testing.md#writing-tests-for-nextjs
- Errors have a helpful link attached, see
https://github.com/vercel/next.js/blob/canary/contributing.md

### Adding a feature

- Implements an existing feature request or RFC. Make sure the feature
request has been accepted for implementation before opening a PR. (A
discussion must be opened, see
https://github.com/vercel/next.js/discussions/new?category=ideas)
- Related issues/discussions are linked using `fixes #number`
- e2e tests added
(https://github.com/vercel/next.js/blob/canary/contributing/core/testing.md#writing-tests-for-nextjs)
- Documentation added
- Telemetry added. In case of a feature if it's used or not.
- Errors have a helpful link attached, see
https://github.com/vercel/next.js/blob/canary/contributing.md


## For Maintainers

- Minimal description (aim for explaining to someone not on the team to
understand the PR)
- When linking to a Slack thread, you might want to share details of the
conclusion
- Link both the Linear (Fixes NEXT-xxx) and the GitHub issues
- Add review comments if necessary to explain to the reviewer the logic
behind a change

### What?

### Why?

### How?

Closes NEXT-
Fixes #

-->